fix: 데이터베이스 마이그레이션에서 테스트 환경에서의 컬럼 제거 및 추가 조건 추가

+++ db/migrate/20250623100402_add_vector_body.rb
@@ -1,10 +1,10 @@
 class AddVectorBody < ActiveRecord::Migration[8.0]
   def up
     enable_extension "vector" unless Rails.env.test?
-    execute "ALTER TABLE articles ADD COLUMN embedding vector(768);"
+    execute "ALTER TABLE articles ADD COLUMN embedding vector(768);" unless Rails.env.test?
   end
 
   def down
-    remove_column :articles, :embedding, :vector
+    remove_column :articles, :embedding, :vector unless Rails.env.test?
   end
 end
