+++ db/migrate/20250725132244_change_embedding_from_article.rb
@@ -1,14 +1,14 @@
 class ChangeEmbeddingFromArticle < ActiveRecord::Migration[8.0]
 def up
     # 기존 컬럼 삭제
-    remove_column :articles, :embedding
+    remove_column :articles, :embedding unless Rails.env.test?
     # 새로운 크기로 컬럼 생성 (예: 1536 차원)
-    execute "ALTER TABLE articles ADD COLUMN embedding vector(1536);"
+    execute "ALTER TABLE articles ADD COLUMN embedding vector(1536);" unless Rails.env.test?
   end
 
   def down
     # 이전 크기로 되돌리기
-    remove_column :articles, :embedding
-    execute "ALTER TABLE articles ADD COLUMN embedding vector(768);"
+    remove_column :articles, :embedding unless Rails.env.test?
+    execute "ALTER TABLE articles ADD COLUMN embedding vector(768);" unless Rails.env.test?
   end
 end
